fix: upgrade dirname failure fix

+++ upload/install/controller/upgrade/upgrade_2.php
@@ -77,7 +77,7 @@ class Upgrade2 extends \Opencart\System\Engine\Controller {
 							$path = '';
 
 							// Must have a path before files can be moved
-							$directories = explode('/', dirname($destination, '/'));
+							$directories = explode('/', dirname($destination));
 
 							foreach ($directories as $directory) {
 								if (!$path) {
@@ -92,7 +92,7 @@ class Upgrade2 extends \Opencart\System\Engine\Controller {
 							}
 
 							// Check if the path is not directory and check there is no existing file
-							if (substr($destination, -1) != '/') {
+							if (substr($destination, -1) != '/' && !is_dir($base . $destination)) {
 								if (is_file($base . $destination)) {
 									unlink($base . $destination);
 								}
